+++ layers/WPSchema/packages/media/src/Component.php
@@ -20,6 +20,7 @@ class Component extends AbstractComponent
     {
         return [
             \PoPSchema\MediaWP\Component::class,
+            \PoPSchema\QueriedObjectWP\Component::class,
         ];
     }