+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/CHANGELOG.md
@@ -46,7 +46,17 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
   - `PostCategory.urlPath: String!`
   - `User.urlPath: String!`
 - Added field arguments to `Root.mediaItems` for filtering results
-- Added field `MediaItem.srcSet`
+- Added fields in `Media`:
+  - `srcSet`
+  - `url`
+  - `urlPath`
+  - `slug`
+  - `title`
+  - `caption`
+  - `altText`
+  - `description`
+  - `date`
+  - `mimeType`
 
 ### Added